Fix solver cascade modeling, par test, and center division; fix compile error

+++ lib/game/physics/soft_body.dart
@@ -132,7 +132,7 @@ class SoftBody {
     for (final p in particles) {
       s += p.pos;
     }
-    return s / particles.length;
+    return s / particles.length.toDouble();
   }
 
   Vec2 get velocity {
